collector/app/requirements.txt
mpipe
dnspython
certstream
cryptography